# gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CareConnect API
  version: 1.0.0
  description: |
    Personal medical clipboard API with PHI isolation, QR consent, and Doctor Finder.

    **Security Model**: Zero-trust, idempotence required for mutations, TLS 1.3.
    **Compliance**: PHI confined to GCP Healthcare API; consent-gated reads.
    **Residency**: Canada (northamerica-northeast1).
  contact:
    name: CareConnect Platform Team
    email: platform@careconnect.health
servers:
  - url: https://api-dev.careconnect.health/v1
    description: Development (CA-dev)
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []
paths:
  /patients/{patientId}/summary:
    get:
      operationId: getPatientSummary
      summary: Get patient summary (clipboard view)
      description: |
        Returns aggregated view: demographics, recent vitals (30d), medications, insurance.
        Requires active consent. Non-PHI synthetic data in demo mode.
      tags:
        - Patients
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Patient summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientSummary'
        '403':
          $ref: '#/components/responses/ConsentRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
        - BearerAuth: []
  /patients/{patientId}/observations:
    get:
      operationId: listObservations
      summary: Get patient observations (vitals, labs)
      description: FHIR Observation resources with filtering. Consent-gated.
      tags:
        - Patients
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/CorrelationId'
        - name: type
          in: query
          description: Observation type (e.g., vital-signs, laboratory)
          schema:
            type: string
        - name: since
          in: query
          description: ISO 8601 date (inclusive)
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          description: ISO 8601 date (exclusive)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Observation list
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Observation'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '403':
          $ref: '#/components/responses/ConsentRequired'
        '404':
          $ref: '#/components/responses/NotFound'
  /consents:
    post:
      operationId: createConsent
      summary: Record patient consent
      description: |
        Create consent record for data sharing. Idempotent.
        Stored as FHIR Consent resource with scope, purpose, expiry.
      tags:
        - Consents
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentCreate'
      responses:
        '201':
          description: Consent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '200':
          description: Idempotent replay (consent already exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT token with tenant/role claims
  parameters:
    PatientId:
      name: patientId
      in: path
      required: true
      description: Patient identifier (pseudonymous)
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID v4 for idempotent mutations
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Request correlation ID for tracing
      schema:
        type: string
  responses:
    ConsentRequired:
      description: Active consent required for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    IdempotencyConflict:
      description: Idempotency-Key reused with different payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    PatientSummary:
      type: object
      required:
        - patientId
        - demographics
        - recentVitals
        - medications
      properties:
        patientId:
          type: string
          format: uuid
        demographics:
          $ref: '#/components/schemas/Demographics'
        healthCard:
          $ref: '#/components/schemas/HealthCard'
        insurance:
          $ref: '#/components/schemas/Insurance'
        recentVitals:
          type: array
          items:
            $ref: '#/components/schemas/Vital'
        medications:
          type: array
          items:
            $ref: '#/components/schemas/Medication'
    Demographics:
      type: object
      required:
        - name
        - dateOfBirth
      properties:
        name:
          type: object
          required:
            - given
            - family
          properties:
            given:
              type: string
            family:
              type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other, unknown]
    HealthCard:
      type: object
      properties:
        number:
          type: string
          description: Provincial health card number (demo only in PWA)
        province:
          type: string
          description: Issuing province (AB, BC, ON, etc.)
        expiryDate:
          type: string
          format: date
    Insurance:
      type: object
      properties:
        provider:
          type: string
        policyNumber:
          type: string
        groupNumber:
          type: string
    Vital:
      type: object
      required:
        - type
        - value
        - unit
        - effectiveDateTime
      properties:
        type:
          type: string
          description: LOINC code or common name
        value:
          type: number
        unit:
          type: string
        effectiveDateTime:
          type: string
          format: date-time
    Medication:
      type: object
      required:
        - name
        - dosage
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        route:
          type: string
    Observation:
      type: object
      required:
        - id
        - type
        - effectiveDateTime
      properties:
        id:
          type: string
        type:
          type: string
        value:
          type: object
        effectiveDateTime:
          type: string
          format: date-time
    ConsentCreate:
      type: object
      required:
        - patientId
        - scope
        - purpose
        - expiresAt
      properties:
        patientId:
          type: string
          format: uuid
        scope:
          type: array
          items:
            type: string
          description: Data categories (e.g., vitals, medications, labs)
        purpose:
          type: string
          description: Use case (e.g., acute-care, research-exempt)
        expiresAt:
          type: string
          format: date-time
        granteeId:
          type: string
          format: uuid
          description: Clinician/organization ID
    Consent:
      allOf:
        - $ref: '#/components/schemas/ConsentCreate'
        - type: object
          required:
            - consentId
            - createdAt
            - status
          properties:
            consentId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            status:
              type: string
              enum: [active, expired, revoked]
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        correlationId:
          type: string
          description: Request correlation ID